# backend/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    name: Delicious Team
  description: |-
    家庭菜谱管理系统 — 支持菜谱版本控制、百科搜索、翻译、导入导出
    所有接口前缀 /api/v1，单用户模式（无需认证）
  title: 人间烟火 (Delicious) API
  version: "2.0"
paths:
...
  /recipes:
    get:
      description: 分页获取当前用户的菜谱列表，支持关键词搜索、评分筛选、排序
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 菜名搜索关键词
        in: query
        name: keyword
        type: string
      - description: 最低评分 (1-5)
        in: query
        name: min_rating
        type: integer
      - description: 最高评分 (1-5)
        in: query
        name: max_rating
        type: integer
      - default: updated_at
        description: 排序字段
        enum:
        - created_at
        - updated_at
        - user_rating
        in: query
        name: order_by
        type: string
      - default: true
        description: 是否降序
        in: query
        name: desc
        type: boolean
      - description: 创建时间起点 (RFC3339)
        in: query
        name: created_after
        type: string
      - description: 创建时间终点 (RFC3339)
        in: query
        name: created_before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: items[], page_info
          schema:
            additionalProperties: true
            type: object
      summary: 菜谱列表
      tags:
      - 菜谱管理
    post:
      consumes:
      - application/json
      description: 新建菜谱并自动创建第一个版本（version 1）
      parameters:
      - description: 菜谱信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: recipe
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建菜谱
      tags:
      - 菜谱管理
  /recipes/{id}:
    delete:
      description: 软删除菜谱，移入回收站。30 天内可通过 /restore 恢复
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 菜谱不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除菜谱（软删除）
      tags:
      - 菜谱管理
    get:
      description: 根据 ID 获取菜谱及当前版本完整信息
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: recipe
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 菜谱不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取菜谱详情
      tags:
      - 菜谱管理
    put:
      consumes:
      - application/json
      description: 编辑菜谱会新增一个不可变版本（版本号+1），不会修改旧版本
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: recipe
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 菜谱不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 编辑菜谱
      tags:
      - 菜谱管理
  /recipes/{id}/diff:
    get:
      description: O(n+m) 算法对比两个版本的配料和步骤差异，返回中文摘要
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 基准版本 ID
        in: query
        name: base_version_id
        required: true
        type: integer
      - description: 对比版本 ID
        in: query
        name: target_version_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: base_version, target_version, diff
          schema:
            additionalProperties: true
            type: object
      summary: 对比两个版本
      tags:
      - 版本管理
  /recipes/{id}/diff/encyclopedia:
    get:
      description: 将菜谱当前版本与关联的百科菜谱进行对比
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 指定百科菜谱 ID（可选，默认用菜谱已关联的）
        in: query
        name: encyclopedia_recipe_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: encyclopedia_recipe_id, encyclopedia_name, encyclopedia_ingredients,
            encyclopedia_process_steps, my_version, diff
          schema:
            additionalProperties: true
            type: object
      summary: 与百科基准对比
      tags:
      - 版本管理
  /recipes/{id}/permanent:
    delete:
      description: 彻底删除菜谱，不可恢复
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 已彻底删除'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 菜谱不在回收站中
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 物理删除
      tags:
      - 回收站
  /recipes/{id}/restore:
    post:
      description: 从回收站恢复已删除的菜谱
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 已恢复'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 菜谱不在回收站中
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 恢复菜谱
      tags:
      - 回收站
  /recipes/{id}/versions:
    get:
      description: 获取某个菜谱的所有历史版本，按版本号降序
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: versions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 菜谱不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 版本历史列表
      tags:
      - 版本管理
  /recipes/{id}/versions/{version_id}:
    get:
      description: 查看某个历史版本的完整配料和步骤
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 版本 ID
        in: path
        name: version_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: version
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 版本不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取某个版本详情
      tags:
      - 版本管理
...
  /recipes/import:
    post:
      consumes:
      - application/json
      description: 从 JSON 导入菜谱。按 name 去重：已存在则追加新版本，不存在则新建
      parameters:
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 导入菜谱
      tags:
      - 导入导出
  /recipes/trash:
    get:
      description: 查看已删除的菜谱
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: items[], page_info
          schema:
            additionalProperties: true
            type: object
      summary: 回收站列表
      tags:
      - 回收站
...
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"